requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
cffi==1.16.0
cryptography==41.0.7
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.4.2
mysqlclient==2.2.4
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.3
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.2.4
typing_extensions==4.8.0
tzdata==2023.3